requirements_minimal.txt
django==4.2.16
python-decouple==3.8
whitenoise==6.6.0
django-allauth==0.57.0
dj-database-url==2.1.0
psycopg2-binary==2.9.11
Pillow==11.3.0